feat: make contact form client-side
